requirements.txt
flask==2.0.1
pillow==9.5.0
tensorflow==2.13.0
numpy>=1.19.2
pathlib==1.0.1
python-dateutil==2.8.2
werkzeug==2.2.3
requests==2.31.0
tzdata>=2023.3
psycopg2-binary==2.9.9
python-dotenv
